feat:認証機能の追加

+++ api/src/index.ts
@@ -2,6 +2,7 @@ import { serve } from '@hono/node-server'
 import { PrismaClient } from '@prisma/client'
 import { Hono } from 'hono'
 import { cors } from 'hono/cors'
+import { clerkMiddleware, getAuth } from '@hono/clerk-auth'
 
 const prisma = new PrismaClient()
 const app = new Hono()
@@ -10,13 +11,23 @@ app.use(
   '/*',
   cors({
     origin: 'http://localhost:5173',
-  })
+    allowHeaders: ['Content-Type', 'Authorization'],
+    allowMethods: ['POST', 'GET', 'PUT', 'DELETE', 'OPTIONS'],
+    credentials: true,
+  }),
 )
+
+app.use('*', clerkMiddleware())
 
 // 1. 全件取得 (GET)
 app.get('/todos', async (c) => {
+  const auth = getAuth(c);
+  if(!auth?.userId) {
+    return c.json({ todos: []});
+  }
   try {
     const todos = await prisma.todo.findMany({
+      where: { userId: auth.userId },
       orderBy: { createdAt: 'desc' }
     });
     return c.json({ todos });
@@ -27,11 +38,15 @@ app.get('/todos', async (c) => {
 
 // 2. 新規作成 (POST)
 app.post('/todos', async (c) => {
+  const auth = getAuth(c)
+  if (!auth?.userId) return c.json({ error: "Unauthorized" }, 401)
+  
   try {
     const { title } = await c.req.json();
     const newTodo = await prisma.todo.create({
       data: {
         title,
+        userId: auth.userId,
         completed: false,
       },
     });
@@ -43,12 +58,18 @@ app.post('/todos', async (c) => {
 
 // 3. 更新 (PUT) - 完了状態の切り替え
 app.put("/todos/:id", async (c) => {
+  const auth = getAuth(c)
+  if (!auth?.userId) return c.json({ error: "Unauthorized" }, 401)
+  
   try {
     const id = Number(c.req.param('id'));
     const { completed } = await c.req.json();
     
     const updatedTodo = await prisma.todo.update({
-      where: { id },
+      where: { 
+        id: id,
+        userId: auth.userId,
+      },
       data: { completed },
     });
     return c.json(updatedTodo);
@@ -59,10 +80,16 @@ app.put("/todos/:id", async (c) => {
 
 // 4. 個別削除 (DELETE)
 app.delete("/todos/:id", async (c) => {
+  const auth = getAuth(c)
+  if (!auth?.userId) return c.json({ error: "Unauthorized" }, 401)
+
   try {
     const id = Number(c.req.param('id'));
     await prisma.todo.delete({
-      where: { id },
+      where: {
+        id: id,
+        userId: auth.userId,
+      },
     });
     return c.json({ message: "Deleted", id });
   } catch (error) {
@@ -72,8 +99,13 @@ app.delete("/todos/:id", async (c) => {
 
 // 5. 全削除 (DELETE)
 app.delete("/todos", async (c) => {
+  const auth = getAuth(c)
+  if (!auth?.userId) return c.json({ error: "Unauthorized" }, 401)
+
   try {
-    await prisma.todo.deleteMany(); // SQL の TRUNCATE または DELETE FROM に相当
+    await prisma.todo.deleteMany({
+      where: { userId: auth.userId }
+    }); // SQL の TRUNCATE または DELETE FROM に相当
     return c.json({ message: "All todos deleted" });
   } catch (error) {
     return c.json({ error: "全削除に失敗しました" }, 500);
